requirements.txt
flasgger==0.9.7.1
Flask==3.1.0
jsonify==0.5
mysql-connector-python==9.2.0